refactor(store-dir): extract inline test modules into dedicated files
Move five inline `#[cfg(test)] mod tests { ... }` blocks (621, 298,
253, 124, and 115 lines respectively) into external `tests.rs` files
following the convention in CODE_STYLE_GUIDE.md. No behavioral change.

Part of #358.
